requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0+cu124
torchvision==0.21.0+cu124
numpy
matplotlib
Pillow
albumentations
tqdm
scikit-learn
